src/django/smart_datalake.tmp/requirements.txt
asgiref
Django
djangorestframework
mysqlclient
sqlparse
typing_extensions
